requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
dropbox>=11.36.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
